:star2: feat: edit blips in memory
