+++ src/components/Radar.vue
@@ -44,6 +44,7 @@
 
 <script>
 import NewBlip from './NewBlip'
+import { getPseudoRand } from '../util'
 
 export default {
   components: { NewBlip },
@@ -73,21 +74,6 @@ export default {
     arrangeBlips () {
       function getDomWidth (domClass) {
         return document.getElementsByClassName(domClass)[0].clientWidth
-      }
-      // generate hash from string
-      function getHash (string) {
-        let h = 0
-        for (const char of string) {
-          h = ((h << 5) - h) + char.charCodeAt(0)
-          h |= 0
-        }
-        return h
-      }
-      // create 0..1 pseudo random from string
-      function getPseudoRand (string) {
-        const h = getHash(string)
-        // convert signed int32 space to 0..1 float
-        return (h + Math.pow(2, 31)) / Math.pow(2, 32)
       }
       const blips = document.getElementsByClassName('blip')
       for (let b of blips) {
